pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "clerk-backend-api>=3.0.5",
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "openai>=1.95.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
